refactor: 拆分 function.ts 与 main.css，src 目录按 core/ui/download/network 分层，硬编码抽取到 data/*.json
