+++ src/core/enum.ts
