+++ src/core/hijack.ts
