+++ src/css/overlay.css
@@ -1,0 +1,69 @@
+#pluginOverlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.75);
+    z-index: 2147483645;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+}
+
+#pluginOverlay .main {
+    color: var(--text);
+    font-size: 24px;
+    width: 60%;
+    background-color: var(--body);
+    padding: 24px;
+    margin: 10px;
+    overflow-y: auto;
+}
+
+@media (max-width: 640px) {
+    #pluginOverlay .main {
+        width: 100%;
+    }
+}
+
+#pluginOverlay button {
+    padding: 10px 20px;
+    color: var(--primary-text);
+    font-size: 18px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+}
+
+#pluginOverlay button {
+    background-color: var(--primary);
+}
+
+#pluginOverlay button[disabled] {
+    background-color: var(--muted);
+    cursor: not-allowed;
+}
+
+#pluginOverlay .checkbox {
+    width: 32px;
+    height: 32px;
+    margin: 0 4px 0 0;
+    padding: 0;
+}
+
+#pluginOverlay .checkbox-container {
+    display: flex;
+    align-items: center;
+    margin: 0 0 10px 0;
+}
+
+#pluginOverlay .checkbox-label {
+    color: var(--text);
+    font-size: 32px;
+    font-weight: bold;
+    margin-left: 10px;
+    display: flex;
+    align-items: center;
+}
