+++ src/data/downloadLinks.json
@@ -1,0 +1,32 @@
+[
+    "iwara.zip",
+    "pan.baidu",
+    "/s/",
+    "mega.nz",
+    "drive.google.com",
+    "aliyundrive",
+    "uploadgig",
+    "katfile",
+    "storex",
+    "subyshare",
+    "rapidgator",
+    "filebe",
+    "filespace",
+    "mexa.sh",
+    "mexashare",
+    "mx-sh.net",
+    "icerbox",
+    "alfafile",
+    "1drv.ms",
+    "onedrive.",
+    "gofile.io",
+    "workupload.com",
+    "pixeldrain.",
+    "dailyuploads.net",
+    "katfile.com",
+    "fikper.com",
+    "frdl.io",
+    "rg.to",
+    "gigafile.nu",
+    "mediafire.com"
+]
