+++ src/ui/toastify.ts
@@ -1,18 +1,19 @@
-import { Dictionary } from "./class";
-import { debounce, isNullOrUndefined, UUID } from "./env";
+import { Dictionary } from "../core/class";
+import { debounce, isNullOrUndefined, UUID } from "../core/env";
 export type Gravity = 'top' | 'bottom';
 export type Position = 'left' | 'center' | 'right';
 export type CloseReason = 'timeout' | 'close-button' | 'other';
 export const activeToasts = new Dictionary<Toast>();
 const toastTimeouts = new Map<Toast, number>();
-const toastIntervals = new Map<Toast, number>();
 const toastContainers = new Map<string, HTMLElement>();
 const offscreenContainer = document.createElement('div')
 offscreenContainer.classList.add('offscreen-container')
+const camelToKebab = (str: string): string => str.replace(/([A-Z])/g, '-$1').toLowerCase()
 const getContainer = (gravity: Gravity, position: Position): HTMLElement => {
     const containerId = `toast-container-${gravity}-${position}`
-    if (!toastContainers.has(containerId)) {
-        const container = document.createElement('div')
+    let container = toastContainers.get(containerId)
+    if (isNullOrUndefined(container)) {
+        container = document.createElement('div')
         container.id = containerId
         container.classList.add(
             'toast-container',
@@ -22,7 +23,7 @@ const getContainer = (gravity: Gravity, position: Position): HTMLElement => {
         document.body.appendChild(container)
         toastContainers.set(containerId, container)
     }
-    return toastContainers.get(containerId)!
+    return container
 }
 const addTimeout = (toast: Toast, callback: () => void): void => {
     if (isNullOrUndefined(toast.options.duration)) return
@@ -36,16 +37,14 @@ const addTimeout = (toast: Toast, callback: () => void): void => {
 
     if (!toast.showProgress) return
     if (isNullOrUndefined(toast.progress)) return
-    const startTime = Date.now()
-    const updateRemainingTime = () => {
-        if (isNullOrUndefined(toast.progress)) return
-        const elapsed = Date.now() - startTime
-        const remaining = Math.max(0, duration - elapsed)
-        toast.progress.style.setProperty('--toast-progress', `${remaining / duration}`)
-    }
-    toast.progress.style.setProperty('--toast-progress', `1`)
-    const intervalId = window.setInterval(updateRemainingTime, 20)
-    toastIntervals.set(toast, intervalId)
+    // 进度条由纯 CSS animation 驱动（替代 20ms setInterval 写 CSS 变量）
+    // duration 变化或重新计时（mouseleave 恢复）时：移除动画 + 强制 reflow 后恢复，从头播放
+    const progress = toast.progress
+    progress.style.setProperty('--toast-duration', `${duration}ms`)
+    progress.style.animation = 'none'
+    void progress.offsetWidth
+    progress.style.animation = ''
+    progress.style.animationPlayState = 'running'
 }
 const delTimeout = (toast: Toast): void => {
     const timeoutId = toastTimeouts.get(toast)
@@ -54,13 +53,9 @@ const delTimeout = (toast: Toast): void => {
         toastTimeouts.delete(toast)
     }
     if (!toast.showProgress) return
-    const intervalId = toastIntervals.get(toast)
-    if (!isNullOrUndefined(intervalId)) {
-        clearInterval(intervalId)
-        toastIntervals.delete(toast)
-    }
+    // 暂停进度条动画（mouseover 暂停 / hide 时停止，不再需要 interval）
     if (!isNullOrUndefined(toast.progress)) {
-        toast.progress.style.removeProperty('--toast-progress')
+        toast.progress.style.animationPlayState = 'paused'
     }
 }
 export interface ToastOptions {
@@ -102,8 +97,7 @@ interface Options {
  * new Toast({ text: 'Hello World' }).show()
  */
 export class Toast {
-    private static readonly defaults: Options = {
-        id: UUID(),
+    private static readonly defaults: Omit<Options, 'id'> = {
         gravity: 'top',
         position: 'left',
         stopOnFocus: true,
@@ -127,16 +121,19 @@ export class Toast {
     private animationEndHandler?: (e: AnimationEvent) => void
     private clickHandler?: (e: MouseEvent) => void
     private closeButton?: HTMLSpanElement
+    private hidden = false
     /**
      * Create a Toastify instance
      * @param options User configuration options
      */
     constructor(options: ToastOptions) {
+        // id 缺省时每次生成新 UUID（避免所有无 id 的 toast 共享 defaults 中的同一个 id）
+        this.id = options.id ?? UUID()
         this.options = {
             ...Toast.defaults,
-            ...options
+            ...options,
+            id: this.id
         }
-        this.id = this.options.id
         this.root = getContainer(this.options.gravity, this.options.position)
         this.gravity = this.options.gravity
         this.position = this.options.position
@@ -227,9 +224,6 @@ export class Toast {
         return this
     }
     private applyStyles(element: HTMLElement, styles: Partial<CSSStyleDeclaration>) {
-        function camelToKebab(str: string): string {
-            return str.replace(/([A-Z])/g, '-$1').toLowerCase()
-        }
         for (const key in styles) {
             const value = styles[key]
             const property = camelToKebab(key)
@@ -252,11 +246,7 @@ export class Toast {
         if (this.oldestFirst) {
             this.root.insertBefore(this.element, this.root.firstChild)
         } else {
-            if (this.root.lastChild) {
-                this.root.insertBefore(this.element, this.root.lastChild.nextSibling)
-            } else {
-                this.root.appendChild(this.element)
-            }
+            this.root.appendChild(this.element)
         }
         return this
     }
@@ -304,19 +294,20 @@ export class Toast {
      * Triggers a CSS exit animation and removes the element after the animation completes
      */
     public hide(reason: CloseReason = 'other'): void {
-        if (!this.element) return
+        // 幂等：重复 hide（如关闭按钮 click 冒泡 + onClick 同时触发）只执行一次，避免 onClose 重复调用
+        if (this.hidden || !this.element) return
+        this.hidden = true
         delTimeout(this)
         activeToasts.delete(this.id)
         this.animationEndHandler = (e: AnimationEvent) => {
             if (e.animationName.startsWith('toast-out')) {
-                this.element.removeEventListener('animationend', this.animationEndHandler!)
                 this.element.remove()
                 this.options.onClose?.call(this, new CustomEvent('toast-close', {
                     detail: { reason }
                 }))
             }
         }
-        this.element.addEventListener('animationend', this.animationEndHandler)
+        this.element.addEventListener('animationend', this.animationEndHandler, { once: true })
         this.removeEventListeners()
             .toggleAnimationState(false)
     }
