+++ test/tests/path.test.ts
@@ -1,6 +1,6 @@
 import '../setup.ts';
 import { Test, TestGroup } from '../framework.ts';
-import { Path } from '../../src/class.ts';
+import { Path } from '../../src/core/class.ts';
 
 const pathTestGroup = new TestGroup('Path 类', '路径解析与规范化测试');
